backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Pydantic for data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# JSON logging
python-json-logger==2.0.7

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# MongoDB
motor==3.3.2
pymongo==4.6.1

# Excel support
openpyxl==3.1.2

# HTTP client
httpx==0.26.0

# Async file operations
aiofiles==23.2.1

# UUID generation
uuid6==2024.1.12

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
isort==5.13.2
mypy==1.8.0
